ErrorProne: IntLongMath

+++ examples/ScrollIndicator/src/java/example/MainPanel.java
@@ -60,7 +60,7 @@ class ScrollIndicator extends JProgressBar {
   }
 
   @Override public double getPercentComplete() {
-    long span = model.getMaximum() - model.getMinimum();
+    int span = model.getMaximum() - model.getMinimum();
     double currentValue = model.getValue() + model.getExtent();
     return (currentValue - model.getMinimum()) / span;
   }
